+++ src/core/mod.rs
@@ -7,4 +7,7 @@ pub mod barcode;
 #[cfg(feature = "decode")]
 pub mod decode;
 
+#[cfg(feature = "decode-any")]
+pub mod decode_any;
+
 pub mod types;